ai-chatbot/backend/requirements.txt
fastapi
uvicorn
langchain
pypdf
faiss-cpu
pymongo
python-dotenv
ollama
sentence-transformers
torch